refactor(frontend): correct PascalCase rename components with QR in the name
