+++ src/frontend/src/lib/components/common/QrButton.svelte
