+++ src/frontend/src/lib/components/receive/ReceiveQrCode.svelte
