+++ src/frontend/src/lib/components/referral/ReferralCodeModal.svelte
@@ -10,7 +10,7 @@
 	import ContentWithToolbar from '$lib/components/ui/ContentWithToolbar.svelte';
 	import ExternalLink from '$lib/components/ui/ExternalLink.svelte';
 	import MessageBox from '$lib/components/ui/MessageBox.svelte';
-	import SkeletonQRCode from '$lib/components/ui/SkeletonQRCode.svelte';
+	import SkeletonQrCode from '$lib/components/ui/SkeletonQrCode.svelte';
 	import SkeletonText from '$lib/components/ui/SkeletonText.svelte';
 	import { OISY_REFERRAL_URL } from '$lib/constants/oisy.constants';
 	import {
@@ -57,7 +57,7 @@
 					</div>
 				</QRCode>
 			{:else}
-				<SkeletonQRCode />
+				<SkeletonQrCode />
 			{/if}
 		</div>
