+++ src/frontend/src/lib/components/send/SendQrCodeScan.svelte
