+++ src/frontend/src/lib/components/ui/SkeletonQrCode.svelte
